@Xor-el 2020-05-30
@azureskydiver , unfortunately this pull request failed too, I think this is due to `using System.Runtime.InteropServices;` missing from the `using` clause, can you please resolve this so I can merge. Thanks.