@azureskydiver 2020-05-30
It could be that, or it could be my messed up local repro. I may have toasted my master branch and then made things worse after that. I'm going to re-fork and start from scratch.